reafactor: 모듈 이동, 함수 추출
- body를 파싱하는 부분을 함수로 추출
